requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
attrs==25.4.0
authlib==1.6.5
beartype==0.22.6
bitsandbytes==0.48.2
brotli==1.2.0
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.4
dataclasses-json==0.6.7
datasets==4.4.1
ddgs==9.9.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
evaluate==0.4.6
exceptiongroup==1.3.0
fake-useragent==2.2.0
fastapi==0.121.3
fastmcp==2.13.1
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.51.0
googleapis-common-protos==1.72.0
gradio==5.50.0
gradio-client==1.14.0
greenlet==3.2.4
groovy==0.1.2
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.76.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-pydantic==0.6
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.0.8
langchain-anthropic==1.1.0
langchain-cerebras==0.7.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-google-genai==3.1.0
langchain-groq==1.0.1
langchain-huggingface==1.0.1
langchain-mcp-adapters==0.1.13
langchain-mcp-tools==0.2.13
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langchain-xai==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.44
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mcp==1.19.0
mcp-chat==0.2.14
mdurl==0.1.2
modal==1.2.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy-extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.5
openai==2.8.1
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
peft==0.18.0
pillow==11.3.0
platformdirs==4.5.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.12.0
pydub==0.25.1
pygments==2.19.2
pyjson5==2.0.0
pyjwt==2.10.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rouge-score==0.1.2
rpds-py==0.29.0
rsa==4.9.1
ruff==0.14.6
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sqlalchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
synchronicity==0.10.4
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
trl==0.25.1
typer==0.20.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
unsloth==2024.8
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
